Refactor IDs to Longs from Ints. - Room uses Longs internally Added Notes and Descriptions fields

+++ app/src/main/java/com/example/android/roomwordnavigation/characters/CharacterListViewModel.kt
@@ -19,7 +19,7 @@ class CharacterListViewModel @Inject constructor(private val repository: ICharac
         repository.insert(characterEntity)
     }
 
-    fun get(characterId: Int) = repository.get(characterId)
+    fun get(characterId: Long) = repository.get(characterId)
 
     fun update(characterEntity: CharacterEntity) = viewModelScope.launch(Dispatchers.IO) {
         repository.update(characterEntity)
