+++ app/src/main/java/com/example/android/roomwordnavigation/data/entities/CharacterEntity.kt
@@ -12,7 +12,7 @@ data class CharacterEntity(
     @ColumnInfo(name = "description") val description: String = "",
     @ColumnInfo(name = "notes") val notes: String = "",
     //@ColumnInfo(name = "avatar") val avatar: String="",
-    @ColumnInfo(name = "id") @PrimaryKey(autoGenerate = true) val id: Int = 0
+    @ColumnInfo(name = "id") @PrimaryKey(autoGenerate = true) val id: Long = 0
 ) {
     companion object {
         val diffCallback = object : DiffUtil.ItemCallback<CharacterEntity>() {
